backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Security & Utilities
python-decouple>=3.8
django-ratelimit>=4.1.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
